requirements.txt
pdfplumber
pdf2image
pillow
pytesseract
openpyxl
click
regex
streamlit
